# Bpf/proxy: less evaluations/allocs for debug prints