--- a/bpf/mock/map.go
+++ b/bpf/mock/map.go
@@ -105,3 +105,41 @@ func NewMockMap(params bpf.MapParameters) *Map {
 }
 
 var _ bpf.Map = (*Map)(nil)
+
+type DummyMap struct{}
+
+func (*DummyMap) GetName() string {
+	return "DummyMap"
+}
+
+func (*DummyMap) Open() error {
+	return nil
+}
+
+func (*DummyMap) EnsureExists() error {
+	return nil
+}
+
+func (*DummyMap) MapFD() bpf.MapFD {
+	return 0
+}
+
+func (*DummyMap) Path() string {
+	return "DummyMap"
+}
+
+func (*DummyMap) Iter(_ bpf.MapIter) error {
+	return nil
+}
+
+func (*DummyMap) Update(k, v []byte) error {
+	return nil
+}
+
+func (*DummyMap) Get(k []byte) ([]byte, error) {
+	return nil, unix.ENOENT
+}
+
+func (*DummyMap) Delete(k []byte) error {
+	return nil
+}