--- a/bpf/proxy/kube-proxy_test.go
+++ b/bpf/proxy/kube-proxy_test.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +build !benchmark
+
 package proxy_test
 
 import (