--- a/bpf/proxy/syncer_bench_test.go
+++ b/bpf/proxy/syncer_bench_test.go
@@ -15,8 +15,6 @@
 // We keep the benchmarks in the proxy package to be able to bench unexported
 // partial functionality.
 
-// +build benchmark
-
 package proxy
 
 import (