--- a/bpf/proxy/syncer_test.go
+++ b/bpf/proxy/syncer_test.go
@@ -25,7 +25,6 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/sys/unix"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	k8sp "k8s.io/kubernetes/pkg/proxy"
@@ -896,46 +895,8 @@ var _ = Describe("BPF Syncer", func() {
 	})
 })
 
-type mockMapDummy struct{}
-
-func (*mockMapDummy) GetName() string {
-	return "mockMapDummy"
-}
-
-func (*mockMapDummy) Open() error {
-	return nil
-}
-
-func (*mockMapDummy) EnsureExists() error {
-	return nil
-}
-
-func (*mockMapDummy) MapFD() bpf.MapFD {
-	return 0
-}
-
-func (*mockMapDummy) Path() string {
-	return "mockMapDummy"
-}
-
-func (*mockMapDummy) Iter(_ bpf.MapIter) error {
-	return nil
-}
-
-func (*mockMapDummy) Update(k, v []byte) error {
-	return nil
-}
-
-func (*mockMapDummy) Get(k []byte) ([]byte, error) {
-	return nil, unix.ENOENT
-}
-
-func (*mockMapDummy) Delete(k []byte) error {
-	return nil
-}
-
 type mockNATMap struct {
-	mockMapDummy
+	mock.DummyMap
 	sync.Mutex
 	m map[nat.FrontendKey]nat.FrontendValue
 }
@@ -1017,7 +978,7 @@ func (m *mockNATMap) Delete(k []byte) error {
 }
 
 type mockNATBackendMap struct {
-	mockMapDummy
+	mock.DummyMap
 	sync.Mutex
 	m map[nat.BackendKey]nat.BackendValue
 }
@@ -1099,7 +1060,7 @@ func (m *mockNATBackendMap) Delete(k []byte) error {
 }
 
 type mockAffinityMap struct {
-	mockMapDummy
+	mock.DummyMap
 	sync.Mutex
 	m map[nat.AffinityKey]nat.AffinityValue
 }