## Description

    bpf/proxy: less evaluations/allocs for debug prints
    
    before:
    
    BenchmarkServiceUpdate/Services_1_Endpoints_1-12                  228688              5391 ns/op            2588 B/op         33 allocs/op
    BenchmarkServiceUpdate/Services_1_Endpoints_10-12                 127449              9369 ns/op            3589 B/op        120 allocs/op
    BenchmarkServiceUpdate/Services_10_Endpoints_1-12                  47924             28154 ns/op           11038 B/op        212 allocs/op
    BenchmarkServiceUpdate/Services_100_Endpoints_1-12                  4546            223610 ns/op           93863 B/op       1634 allocs/op
    BenchmarkServiceUpdate/Services_1000_Endpoints_1-12                  481           2422507 ns/op         1286829 B/op      16089 allocs/op
    BenchmarkServiceUpdate/Services_10000_Endpoints_1-12                  48          26761813 ns/op        10815265 B/op     160607 allocs/op
    
    after:
    
    BenchmarkServiceUpdate/Services_1_Endpoints_1-12                  218082              5357 ns/op            2357 B/op         24 allocs/op
    BenchmarkServiceUpdate/Services_1_Endpoints_10-12                 129049              9624 ns/op            3131 B/op         82 allocs/op
    BenchmarkServiceUpdate/Services_10_Endpoints_1-12                  51529             26160 ns/op            8732 B/op        135 allocs/op
    BenchmarkServiceUpdate/Services_100_Endpoints_1-12                  6537            205158 ns/op           76174 B/op       1031 allocs/op
    BenchmarkServiceUpdate/Services_1000_Endpoints_1-12                  531           2175678 ns/op         1111050 B/op      10087 allocs/op
    BenchmarkServiceUpdate/Services_10000_Endpoints_1-12                  52          22809532 ns/op         9055803 B/op     100604 allocs/op


<!-- A few sentences describing the overall goals of the pull request's commits.
Please include
- the type of fix - (e.g. bug fix, new feature, documentation)
- some details on _why_ this PR should be merged
- the details of the testing you've done on it (both manual and automated)
- which components are affected by this PR
-->

## Todos
- [ ] Unit tests (full coverage)
- [ ] Integration tests (delete as appropriate) In plan/Not needed/Done
- [ ] Documentation
- [ ] Backport
- [ ] Release note

## Release Note
<!-- Writing a release note:
- By default, no release note action is required.
- If you're unsure whether or not your PR needs a note, ask your reviewer for guidance.
- If this PR requires a release note, update the block below to include a concise note describing
  the change and any important impacts this PR may have.
-->

```release-note
None required
```