@tomastigera 2020-08-06
> LGTM abaprt from the !benchmark build tags; not sure if that's a normal thing to do in go? Might bite us if we start running benchmarks as part of the normal test runs.

I wasnt happy with that either, finally found why the kube proxies were starting from ginkgo and removed the tag.